+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 Fixed
 ^^^^^
 - Fixed a deadlock in three level nested transactions (#1810)
+- Fix backward_relations in PydanticMeta (#1814)
 
 
 0.22.2